Move load_image to a utils loc
